Robustify `SetupPyIntegrationTest`.
Previously, when the pants `VERSION` was bumped and before publishing
the `test_setup_py_unregistered_pants_plugin` test would fail to resolve
`pantsbuild.pants` at the new version on pypi. Update the test to use a
custom-produced `pantsbuild.pants` distribution and share this logic
with `PantsRequirementIntegrationTest`.

Similar to #5519
