+++ tests/python/pants_test/backend/python/BUILD
@@ -46,10 +46,19 @@ python_tests(
   name='integration',
   sources=globs('*_integration.py'),
   dependencies=[
+    ':pants_requirement_integration_test_base',
+    'src/python/pants/base:build_environment',
+  ],
+  tags={'integration'},
+)
+
+python_library(
+  name='pants_requirement_integration_test_base',
+  sources=['pants_requirement_integration_test_base.py'],
+  dependencies=[
     'src/python/pants/base:build_environment',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test:int-test',
-  ],
-  tags={'integration'},
+  ]
 )
